refactor: extract horizontal and vertical week view data logic into dedicated hooks
